Added initdata function for rooms and bookings. Modified dictionary data structure to work with room class. TO-DO: Modify BST data structure to work with booking class

+++ DSA_Assignment/Dictionary.h
@@ -37,7 +37,7 @@ public:
 	// destructor
 	~Dictionary();
 
-	int64_t hash(KeyType key);
+	int hash(KeyType key);
 
 	// add a new item with the specified key to the Dictionary
 	// pre : none
